Move PollOpt to the common crate
