+++ src/event_imp.rs
@@ -1,6 +1,6 @@
-use std::{fmt, io, ops};
+use std::io;
 
-use crate::{Ready, Interests, Registry, Token};
+use crate::{Interests, PollOpt, Ready, Registry, Token};
 
 /// A value that may be registered with `Registry`
 ///
@@ -245,324 +245,6 @@ impl<T: Evented> Evented for ::std::sync::Arc<T> {
     }
 }
 
-/// Options supplied when registering an `Evented` handle with `Poll`
-///
-/// `PollOpt` values can be combined together using the various bitwise
-/// operators.
-///
-/// For high level documentation on polling and poll options, see [`Poll`].
-///
-/// # Examples
-///
-/// ```
-/// use mio::PollOpt;
-///
-/// let opts = PollOpt::edge() | PollOpt::oneshot();
-///
-/// assert!(opts.is_edge());
-/// assert!(opts.is_oneshot());
-/// assert!(!opts.is_level());
-/// ```
-///
-/// [`Poll`]: struct.Poll.html
-#[derive(Copy, PartialEq, Eq, Clone, PartialOrd, Ord)]
-pub struct PollOpt(usize);
-
-impl PollOpt {
-    /// Return a `PollOpt` representing no set options.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::empty();
-    ///
-    /// assert!(!opt.is_level());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn empty() -> PollOpt {
-        PollOpt(0)
-    }
-
-    /// Return a `PollOpt` representing edge-triggered notifications.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::edge();
-    ///
-    /// assert!(opt.is_edge());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn edge() -> PollOpt {
-        PollOpt(0b0001)
-    }
-
-    /// Return a `PollOpt` representing level-triggered notifications.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::level();
-    ///
-    /// assert!(opt.is_level());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn level() -> PollOpt {
-        PollOpt(0b0010)
-    }
-
-    /// Return a `PollOpt` representing oneshot notifications.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::oneshot();
-    ///
-    /// assert!(opt.is_oneshot());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn oneshot() -> PollOpt {
-        PollOpt(0b0100)
-    }
-
-    /// Returns true if the options include edge-triggered notifications.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::edge();
-    ///
-    /// assert!(opt.is_edge());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn is_edge(&self) -> bool {
-        self.contains(PollOpt::edge())
-    }
-
-    /// Returns true if the options include level-triggered notifications.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::level();
-    ///
-    /// assert!(opt.is_level());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn is_level(&self) -> bool {
-        self.contains(PollOpt::level())
-    }
-
-    /// Returns true if the options includes oneshot.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::oneshot();
-    ///
-    /// assert!(opt.is_oneshot());
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn is_oneshot(&self) -> bool {
-        self.contains(PollOpt::oneshot())
-    }
-
-    /// Returns true if `self` is a superset of `other`.
-    ///
-    /// `other` may represent more than one option, in which case the function
-    /// only returns true if `self` contains all of the options specified in
-    /// `other`.
-    ///
-    /// See [`Poll`] for more documentation on polling.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::oneshot();
-    ///
-    /// assert!(opt.contains(PollOpt::oneshot()));
-    /// assert!(!opt.contains(PollOpt::edge()));
-    /// ```
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::oneshot() | PollOpt::edge();
-    ///
-    /// assert!(opt.contains(PollOpt::oneshot()));
-    /// assert!(opt.contains(PollOpt::edge()));
-    /// ```
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let opt = PollOpt::oneshot() | PollOpt::edge();
-    ///
-    /// assert!(!PollOpt::oneshot().contains(opt));
-    /// assert!(opt.contains(opt));
-    /// assert!((opt | PollOpt::level()).contains(opt));
-    /// ```
-    ///
-    /// [`Poll`]: struct.Poll.html
-    #[inline]
-    pub fn contains(&self, other: PollOpt) -> bool {
-        (*self & other) == other
-    }
-
-    /// Adds all options represented by `other` into `self`.
-    ///
-    /// This is equivalent to `*self = *self | other`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let mut opt = PollOpt::empty();
-    /// opt.insert(PollOpt::oneshot());
-    ///
-    /// assert!(opt.is_oneshot());
-    /// ```
-    #[inline]
-    pub fn insert(&mut self, other: PollOpt) {
-        self.0 |= other.0;
-    }
-
-    /// Removes all options represented by `other` from `self`.
-    ///
-    /// This is equivalent to `*self = *self & !other`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use mio::PollOpt;
-    ///
-    /// let mut opt = PollOpt::oneshot();
-    /// opt.remove(PollOpt::oneshot());
-    ///
-    /// assert!(!opt.is_oneshot());
-    /// ```
-    #[inline]
-    pub fn remove(&mut self, other: PollOpt) {
-        self.0 &= !other.0;
-    }
-}
-
-impl ops::BitOr for PollOpt {
-    type Output = PollOpt;
-
-    #[inline]
-    fn bitor(self, other: PollOpt) -> PollOpt {
-        PollOpt(self.0 | other.0)
-    }
-}
-
-impl ops::BitXor for PollOpt {
-    type Output = PollOpt;
-
-    #[inline]
-    fn bitxor(self, other: PollOpt) -> PollOpt {
-        PollOpt(self.0 ^ other.0)
-    }
-}
-
-impl ops::BitAnd for PollOpt {
-    type Output = PollOpt;
-
-    #[inline]
-    fn bitand(self, other: PollOpt) -> PollOpt {
-        PollOpt(self.0 & other.0)
-    }
-}
-
-impl ops::Sub for PollOpt {
-    type Output = PollOpt;
-
-    #[inline]
-    fn sub(self, other: PollOpt) -> PollOpt {
-        PollOpt(self.0 & !other.0)
-    }
-}
-
-impl fmt::Debug for PollOpt {
-    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut one = false;
-        let flags = [
-            (PollOpt::edge(), "Edge-Triggered"),
-            (PollOpt::level(), "Level-Triggered"),
-            (PollOpt::oneshot(), "OneShot"),
-        ];
-
-        for &(flag, msg) in &flags {
-            if self.contains(flag) {
-                if one {
-                    write!(fmt, " | ")?
-                }
-                write!(fmt, "{}", msg)?;
-
-                one = true
-            }
-        }
-
-        if !one {
-            fmt.write_str("(empty)")?;
-        }
-
-        Ok(())
-    }
-}
-
-#[test]
-fn test_debug_pollopt() {
-    assert_eq!("(empty)", format!("{:?}", PollOpt::empty()));
-    assert_eq!("Edge-Triggered", format!("{:?}", PollOpt::edge()));
-    assert_eq!("Level-Triggered", format!("{:?}", PollOpt::level()));
-    assert_eq!("OneShot", format!("{:?}", PollOpt::oneshot()));
-}
 
 /// An readiness event returned by [`Poll::poll`].
 ///
@@ -658,7 +340,7 @@ pub fn ready_as_usize(events: Ready) -> usize {
 }
 
 pub fn opt_as_usize(opt: PollOpt) -> usize {
-    opt.0
+    opt.as_usize()
 }
 
 pub fn ready_from_usize(events: usize) -> Ready {
@@ -666,7 +348,7 @@ pub fn ready_from_usize(events: usize) -> Ready {
 }
 
 pub fn opt_from_usize(opt: usize) -> PollOpt {
-    PollOpt(opt)
+    PollOpt::from_usize(opt)
 }
 
 // Used internally to mutate an `Event` in place
